Update to Lunr 2.0.1 PMT#: 110757
This commit moves the Lunr index building functionality out of a manager
class into main.  The index object in Lunr 2.x is immutable, and so the
index needs to be built when its initialized. Moving this functionality
to main.js permits passing the index to the search functions that
require it.