Replace 'Category' PMT# 110782
This commit drops the 'Category' from the event object and renames
'Group-Specific' to 'Category'.

The initial 'Category' from Bedeworks is meant University-wide and it
has a confusing meaning within the context of the CTL site.
'Group-Specific' was renamed to 'Category' because its one of the few
fields CTL staff can manipulate in Bedeworks. This will allow staff to
more flexibly label events.