Switch over to CTL proxy
This commit switched over to a CTL hosted cache of the calendar JSON.
The previous proxy we've been using, from LITO, re-mapped some
fields from Bedeworks.  This required updating the event constructor to
match, as well as corresponding tests.

Also, this commit removes a redundant function that would return a date
object for an event start date.