Filter events after a given date.
This commit introduces the feature to filter out date after a given
date, upstream from the existing date filering functionality.

Why was this needed?
We switched to using our own JSON proxy for event data.  We found that
when the proxy fails to update, events in the past would appear in the
calendar.

There's alredy a date filter, why another one?
It made sense to add it 'upstream' at the point where event objects are
instatiated so that it wouldn't conflict with the user facing filtering.
It would be possible to set an initial date filter using the current
functions, but if a user reset the filters, old events would appear.