Refactor validate function
This commit removes kind of kludgey date comparison that depended on
milliseconds from the epoch. This switches back to using the comparitor
of the Date object. It standarizes the the times of of the Date objects
passed in: start dates are set for midnight of that day, and end dates
are set for 23:59 of that day.