Further refactoring of the filterEvents function.
This commit reduces unessecary logic flow in the filterEvents function.
- It consistently relies on using null values to determine if a variable
has been set. Previously, when initializing the drop downs, the default
was "null" as a string. No mas.
- It simplified the logic of determining when the filterEvents function
has been called with arguments or with URL params
- It improves test coverage for various search cases, including
corresponding tests for when search params are passed with the function
call and when picked up from the query string.