Patch bug reading out of Lunr index PMT# 112228
The searchEvents function uses an array index to fetch results out of
the result which are derived from a list of all events. There exists a
mismatch between the array of events that have been filtered down and
the set of all events.

This patch moves the text filter to be first among the filters so the
array indices match up. Its a temp fix until the searchEvents function
is debugged.