Add environmental make targets
I was running into an issue with getting the correct URL to load for PNG
sprites being used by jQuery UI. The default behavior of webpack's
file-loader is to rename the file <md5 hash>.<ext>, and emit the file to
which ever directory is specified as the build dir, often 'dist'.

This gets borked locally because the file is in 'dist/file.png', but
Webpack is writing the filename as 'file.png'. This was easily solved by
setting the 'publicPath' option in the file-loader to 'dist/'.

That's all great and wonderful except that it breaks when its actually
deployed. To fix this, I made use of Webpack's ability to pass in
'environmental' variables, though really 'env' is just an object passed
in from the command, its not a real ENV.

This commit makes use of this to set 'env.production' and 'env.stage'
vars, and conditionally set the value of 'publicPath' based on those
Booles.