CW-552 HTML parsing bug
This commit fixes a bug likely introduced by Bedeworks escaping emitted
HTML. This commit parses the HTML, and then parses it into a document
fragment.

It also fixes an issue where the 'lede' and 'more' text was
was being split on the first appearance of a '.'. This now parses it
into a document fragment so the markup can be handled as DOM nodes. The
lede is now the first child node, and 'more' are the remaining nodes.