Refactor image loading
This commit replaces 'file-loader' with 'url-loader'. This will replace
urls in css files with inlined data urls. These urls were breaking in
production because file-loader was not prepending the correct url. The
images being inlined are small UI assets, so there shouldn't be too
large of a file size penalty.

Because the images are inlined, this also offered a chance to simplify
the build, and remove unnecessary make targets.